+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - Feat: add `ILexer.lex.peek`.
   - Feat: typed lexer. [#17](https://github.com/DiscreteTom/retsac/issues/17)
   - Feat: add `ReadonlyILexer`.
+  - Fix: actions won't be merged into one when calling `Builder.define/anonymous/ignore`. This will fix expectational lexing behaviour.
 - Parser
   - Feat: add `name` for `ASTNode` and `ASTObj`.
     - **_Breaking Change_**: `ASTNode.toTreeString` will also print `ASTNode.name`.